+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidytable 0.9.2 (in development)
 
 #### New functions
+* `cross_join()`
 * `group_cols()`
 * `map_vec()`
 * `pick()`
@@ -14,6 +15,9 @@
 * `separate()`: Can overwrite separated column without removal (#680)
 * `na_if()`: Properly replaces `NA`s when `y` is a vector (#689)
 * `across()`: Anonymous functions are properly translated (#699)
+
+#### Deprecations
+* Using `by = character()` in joins is now deprecated. Users should instead use `cross_join()`.
 
 # tidytable 0.9.1