+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidytable 0.9.2 (in development)
+# tidytable 0.9.2
 
 #### New functions
 * `cross_join()`